third_party/blink/web_tests/external/wpt/xhr/send-redirect-to-cors-expected.txt

This is a testharness.js-based test.
[FAIL] XMLHttpRequest: send() - Redirect to CORS-enabled resource (303 FOO with string and explicit Content-Type safelisted)
  assert_equals: expected "NO" but got "application/x-pony"
[FAIL] XMLHttpRequest: send() - Redirect to CORS-enabled resource (301 POST with string and explicit Content-Type)
  assert_equals: expected 200 but got 0
[FAIL] XMLHttpRequest: send() - Redirect to CORS-enabled resource (301 POST with string and explicit Content-Type safelisted)
  assert_equals: expected "NO" but got "application/x-pony"
[FAIL] XMLHttpRequest: send() - Redirect to CORS-enabled resource (302 POST with string and explicit Content-Type)
  assert_equals: expected 200 but got 0
Harness: the test ran to completion.
